Fix database schema validation errors - add missing columns to users table

+++ src/main/java/com/mahiberawi/dto/group/GroupRequest.java
@@ -7,6 +7,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class GroupRequest {
     @NotBlank(message = "Group name is required")
@@ -33,6 +35,6 @@ public class GroupRequest {
         private Boolean allowMessagePosting = true;
         private Boolean paymentRequired = false;
         private Boolean requireApproval = false;
-        private Double monthlyDues;
+        private BigDecimal monthlyDues;
     }
 }
