+++ src/main/java/com/mahiberawi/dto/membership/MembershipRequest.java
@@ -6,6 +6,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -23,7 +24,7 @@ public class MembershipRequest {
     private LocalDateTime endDate;
 
     @Positive(message = "Fee must be positive")
-    private Double fee;
+    private BigDecimal fee;
 
     private String description;
 }
